Change default normalize from 16 to 1.
When experimenting with a wide range of feature extraction libraries, I
found that supplying normalize=16 (instead of normalize=1) for ASR on
WSJ resulted in vastly different features from Kaldi. Moreover, the
training was much worse. This normalize value is incredibly easy to
overlook, so I suggest that either the default value is changed (which I
do here), or it is explicitly included as an argument to pass in asr.sh
and tts.sh in egs2/TEMPLATE.